// uldb/ngx-unity/src/app/united-cloud/public-cloud/public-cloud-oci/public-cloud-oci-overview/public-cloud-oci-storage-services/oci-buckets.type.ts
export interface PublicCloudOciBuckets {
    name: string;
    compartment_id: string;
    defined_tags: null;
    namespace: string;
    freeform_tags: null;
    created_by: string;
    etag: string;
    time_created: string;
}

export interface PublicCloudOCIBucketFile {
    name: string;
    archival_state: null;
    time_created: string;
    etag: null;
    time_modified: string;
    storage_tier: null;
    md5: null;
    size: null;
}